+++ app/Http/Controllers/DitreskrimumController.php
@@ -105,10 +105,14 @@ class DitreskrimumController extends Controller
             Storage::delete($ditreskrimum->gambar1);
             $data['gambar1'] = $request->file('gambar1')->store('ditreskrimum');
         } else if ($request->hasFile('gambar2')) {
-            Storage::delete($ditreskrimum->gambar2);
+            if ($ditreskrimum->gambar2) {
+                Storage::delete($ditreskrimum->gambar2);
+            }
             $data['gambar2'] = $request->file('gambar2')->store('ditreskrimum');
         } else if ($request->hasFile('gambar3')) {
-            Storage::delete($ditreskrimum->gambar3);
+            if ($ditreskrimum->gambar3) {
+                Storage::delete($ditreskrimum->gambar3);
+            }
             $data['gambar3'] = $request->file('gambar3')->store('ditreskrimum');
         }
